FIX: Move out of 'fmu' namespace
